Fixed a bug that caused the Node to occasionally get into a loop during the cleanup method.
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/DJob/trunk@4824 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ DistribJob/lib/perl/Node.pm
@@ -338,11 +338,11 @@ sub cleanUp {
 
   print "Cleaning up node $self->{nodeNum}...\n";
 
+  ##now call the task->cleanUpNode method to enable  users to stop processes running on node
   my $task = $self->getTask();
   $task->cleanUpNode($self) if $task;
 
-  if($self->getPort()){
-    ##now call the task->cleanUpNode method to enable  users to stop processes running on node
+  if($self->{nodeNum} && $self->getPort()){
     $self->runCmd("/bin/rm -r $self->{nodeDir}", 1);
     $self->runCmd("closeAndExit",1);
     $self->closePort();
